Create separate files for UploadOptions, UploadErrors and String extension.
